backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.3.7
httpx==0.25.2
# Authentication dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0